+++ lib/sucker_punch/testing/inline.rb
@@ -5,4 +5,3 @@ module Celluloid
     end
   end
 end
-